python-fastapi-spa/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
pyyaml
black
orjson
pytest
httpx
faker
